+++ cmd/root.go
@@ -1,13 +1,12 @@
 package cmd
 
 import (
-	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
 
-	"github.com/ubclaunchpad/inertia/cfg"
 	inertiacmd "github.com/ubclaunchpad/inertia/cmd/cmd"
+	configcmd "github.com/ubclaunchpad/inertia/cmd/config"
 	hostcmd "github.com/ubclaunchpad/inertia/cmd/host"
 	"github.com/ubclaunchpad/inertia/cmd/inpututil"
 	"github.com/ubclaunchpad/inertia/cmd/printutil"
@@ -60,9 +59,7 @@ Issue tracker: https://github.com/ubclaunchpad/inertia/issues`,
 
 	// add children
 	newInitCmd(root)
-	newResetCmd(root)
-	newSetCmd(root)
-	newUpgradeCmd(root)
+	configcmd.AttachConfigCmd(root)
 	remotecmd.AttachRemoteCmd(root)
 	provisioncmd.AttachProvisionCmd(root)
 	hostcmd.AttachHostCmds(root)
@@ -127,83 +124,4 @@ func newInitCmd(inertia *inertiacmd.Cmd) {
 	}
 	init.Flags().String("version", inertia.Version, "specify Inertia daemon version to use")
 	inertia.AddCommand(init)
-}
-
-func newResetCmd(inertia *inertiacmd.Cmd) {
-	var reset = &cobra.Command{
-		Use:   "reset",
-		Short: "Remove inertia configuration from this repository",
-		Long:  `Removes Inertia configuration files pertaining to this project.`,
-		Run: func(cmd *cobra.Command, args []string) {
-			println("WARNING: This will remove your current Inertia configuration")
-			println("and is irreversible. Continue? (y/n)")
-			var response string
-			_, err := fmt.Scanln(&response)
-			if err != nil {
-				printutil.Fatal("invalid response - aborting")
-			}
-			if response != "y" {
-				printutil.Fatal("aborting")
-			}
-			path, err := common.GetFullPath(inertia.ConfigPath)
-			if err != nil {
-				printutil.Fatal(err)
-			}
-			os.Remove(path)
-			println("Inertia configuration removed.")
-		},
-	}
-	inertia.AddCommand(reset)
-}
-
-func newSetCmd(inertia *inertiacmd.Cmd) {
-	var set = &cobra.Command{
-		Use:   "set [property] [value]",
-		Short: "Update a property of your Inertia project configuration",
-		Long:  `Updates a property of your Inertia project configuration and save it to inertia.toml.`,
-		Args:  cobra.ExactArgs(2),
-		Run: func(cmd *cobra.Command, args []string) {
-			// Ensure project initialized.
-			config, path, err := local.GetProjectConfigFromDisk(inertia.ConfigPath)
-			if err != nil {
-				printutil.Fatal(err)
-			}
-			success := cfg.SetProperty(args[0], args[1], config)
-			if success {
-				config.Write(path)
-				println("Configuration setting '" + args[0] + "' has been updated..")
-			} else {
-				println("Configuration setting '" + args[0] + "' not found.")
-			}
-		},
-	}
-	inertia.AddCommand(set)
-}
-
-func newUpgradeCmd(inertia *inertiacmd.Cmd) {
-	var upgrade = &cobra.Command{
-		Use:   "upgrade",
-		Short: "Upgrade your Inertia configuration version to match the CLI",
-		Long:  `Upgrade your Inertia configuration version to match the CLI and saves it to inertia.toml`,
-		Run: func(cmd *cobra.Command, args []string) {
-			// Ensure project initialized.
-			config, path, err := local.GetProjectConfigFromDisk(inertia.ConfigPath)
-			if err != nil {
-				printutil.Fatal(err)
-			}
-
-			var version = inertia.Version
-			if v, _ := cmd.Flags().GetString("version"); v != "" {
-				version = v
-			}
-
-			fmt.Printf("Setting Inertia config to version '%s'", version)
-			config.Version = version
-			if err = config.Write(path); err != nil {
-				printutil.Fatal(err)
-			}
-		},
-	}
-	upgrade.Flags().String("version", inertia.Version, "specify Inertia daemon version to set")
-	inertia.AddCommand(upgrade)
 }
